ldir_core::layout

Module linebreak

Line-breaking algorithm modules.

Modules

cjk

CJK-aware line breaking.

Structs

FitnessClass

Fitness class of a line (Knuth-Plass classification).

LineBreakItem

A single item to be placed by the line-breaking algorithm.

LineBreakOptions

Options controlling the line-breaking algorithm.

LineBreakResult

Result of the line-breaking algorithm.

Functions

compute_adjustment_ratio

Compute the adjustment ratio for a line.

compute_badness

Compute the badness of a line given its adjustment ratio.

compute_demerits

Compute the demerits for a single line break.

compute_demerits_batch

Compute demerits for 4 candidate breaks simultaneously.

compute_demerits_batch_simd⚠

SIMD-accelerated demerits computation for a batch of 4 candidates.

linebreak

Run the Knuth-Plass line-breaking algorithm.